moved lookup user classes to endpoints.lookupuser
