+++ src/main/java/com/photosharesite/backend/endpoints/lookupuser/LookupUserRequest.java
@@ -1,4 +1,4 @@
-package com.photosharesite.backend.api;
+package com.photosharesite.backend.endpoints.lookupuser;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
